1.1: astyle --style=allman
