+++ readme.txt
@@ -3,7 +3,7 @@ HOW TO USE
 1. Create PDF 'document.pdf' from all JPEG files in current directory:
 
     jpeg2pdf *.jpg -o document.pdf
-	
+
 
 2. Create PDF 'scans.pdf' with:
    page size determined by maximum image size,
@@ -21,28 +21,19 @@ HOW TO USE
    crop the height of each page
    (works well for properly cropped photos of A5 pages in landscape
     with unknown/wrong dpi and approximately correct widths)
-   
+
     jpeg2pdf -o from_photos.pdf -p A5 -n landscape -z fw -r height *.jpg
-	
+
 4. It is possible to specify the title, author, subject, keyword, and creator
    of produced PDF document with switches -t, -a, -s, -k, -c:
 
     jpeg2pdf -o doc.pdf -t "The Title of PDF" -a Andrew -c jpeg2pdf *.jpg
-	
-	
+
+
 HOW TO COMPILE AND INSTALL
 
-On Linux:
-
    'make && make install' compiles and installs jpeg2pdf to /usr/local/bin.
-	
-On Windows:
 
-   'buildit.bat' compiles the program, you'll need to copy jpeg2pdf.exe
-   to some directory listed in the PATH environment variable to be able
-   to run it from arbitrary location.
-
-	
 PROJECT HOME
 
    http://sourceforge.net/projects/jpeg2pdf/
